requirements.txt
uagents>=0.23.1
python-dotenv>=1.0.1
supabase>=2.7.0
anthropic>=0.34.0
twilio>=9.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
